Refactor isValidLCDate function into utils folder to prevent index.js running during tests
