+++ index.js
@@ -1,11 +1,7 @@
 const TelegramBot = require('node-telegram-bot-api');
 const axios = require('axios');
 const { Client } = require('pg');
-const dayjs = require('dayjs');
-let isBetween = require('dayjs/plugin/isBetween');
-let utc = require('dayjs/plugin/utc');
-dayjs.extend(isBetween);
-dayjs.extend(utc);
+const isValidLCDate = require('./utils/isValidLCDate');
 const cron = require('node-cron');
 const client = new Client();
 
@@ -238,28 +234,6 @@ bot.onText(
     });
   }
 );
-
-function isValidLCDate(...args) {
-  let submissionHour;
-  let leftBound, rightBound;
-  const resp = args[0];
-  if (args.length > 1) {
-    leftBound = dayjs.utc(args[1], 'YYYYMMDD').hour(0).minute(0).second(0);
-    rightBound = leftBound.add(1, 'day');
-    submissionHour = args[2];
-  } else {
-    leftBound = dayjs.utc().hour(0).minute(0).second(0).millisecond(0);
-    rightBound = leftBound.add(1, 'day');
-    submissionHour = dayjs.utc().hour();
-  }
-
-  const submissionDate = dayjs.utc(resp, 'YYYYMMDD').hour(submissionHour);
-  //  console.log(`Submission hour: ${submissionHour}`);
-  //  console.log(`Submission date: ${submissionDate}`);
-  //  console.log(`Left bound: ${leftBound}`);
-  //  console.log(`Right bound: ${rightBound}`);
-  return submissionDate.isBetween(leftBound, rightBound, 'millisecond', '[)');
-}
 
 // motivational reply to encourage ppl to carry on joining the LC party
 bot.on('message', async (msg) => {
